# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GGC Result System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /results/course/{courseId}:
    get:
      operationId: getCourseResults
      tags: [results]
      summary: Get all student results for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentResult"
  /results/student:
    get:
      operationId: getStudentResults
      tags: [results]
      summary: Lookup results by roll number and session
      parameters:
        - name: rollNo
          in: query
          required: true
          schema:
            type: string
        - name: session
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Student results across all courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentLookupResult"
  /results/transcript/{rollNo}:
    get:
      operationId: downloadTranscript
      tags: [results]
      summary: Download PDF transcript for a student
      parameters:
        - name: rollNo
          in: path
          required: true
          schema:
            type: string
        - name: session
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary

  # ─── Analytics ────────────────────────────────────────────────
...
components:
...
  schemas:
...
    StudentResult:
      type: object
      properties:
        id:
          type: integer
        rollNo:
          type: string
        name:
          type: string
        fatherName:
          type: string
        cnic:
          type: string
        session:
          type: string
        internalMarks:
          type: number
        midTerm:
          type: number
        finalTerm:
          type: number
        practicalWork:
          type: number
        totalObtained:
          type: number
        percentage:
          type: number
        grade:
          type: string
        gradePoint:
          type: number
        status:
          type: string
      required: [id, rollNo, name, percentage, grade, status]
    StudentLookupResult:
      type: object
      properties:
        student:
          type: object
          properties:
            rollNo:
              type: string
            name:
              type: string
            fatherName:
              type: string
            cnic:
              type: string
            session:
              type: string
            departmentName:
              type: ["string", "null"]
          required: [rollNo, name]
        results:
          type: array
          items:
            $ref: "#/components/schemas/StudentResult"
        cgpa:
          type: number
        totalCredits:
          type: number
        passCount:
          type: integer
        failCount:
          type: integer
      required: [student, results, cgpa]
...